added shardSessionLock to prevent concurrent write panic - added test
